+++ packages/monitoring/src/next.ts
@@ -1,0 +1,6 @@
+import * as Sentry from "@sentry/nextjs"
+
+const {withSentry, withSentryConfig} = Sentry
+
+export default Sentry
+export {withSentry, withSentryConfig}